+++ app/controllers/invites_controller.rb
@@ -2,7 +2,11 @@ class InvitesController < ApplicationController
 
   before_action :authenticate_user!
 
-  def new
+  def members
+    @invite = Invite.new
+  end
+
+  def providers
     @invite = Invite.new
   end
 
@@ -10,10 +14,15 @@ class InvitesController < ApplicationController
     @invite = Invite.where(user_id: current_user.id).new(invite_params)
     if @invite.save
       if @invite.recipient_id.present?
-        redirect_to new_invite_path, warning: 'Thank you, but they have already joined!'
+        redirect_to members_invites_path, warning: 'Thank you, but they have already joined!'
       else
-        InviteMailer.new_user_invite(@invite, current_user).deliver_later
-        redirect_to new_invite_path, notice: 'Thank you!'
+        if @invite.for_provider?
+          InviteMailer.new_provider_invite(@invite, current_user).deliver_later
+          redirect_to providers_invites_path, notice: 'Thank you!'
+        else
+          InviteMailer.new_member_invite(@invite, current_user).deliver_later
+          redirect_to members_invites_path, notice: 'Thank you!'
+        end
       end
     end
   end
@@ -21,7 +30,7 @@ class InvitesController < ApplicationController
   private
 
     def invite_params
-      params.require(:invite).permit(:email)
+      params.require(:invite).permit(:email, :for_provider)
     end
 
 end