+++ db/migrate/20150617181600_add_for_provider_to_invites.rb
@@ -1,0 +1,5 @@
+class AddForProviderToInvites < ActiveRecord::Migration
+  def change
+    add_column :invites, :for_provider, :boolean, null: false, default: false
+  end
+end