+++ CHANGELOG.md
@@ -22,7 +22,8 @@
   - Moved ISSS from learn page to partners page
   - Updated link to MyApnea.Org welcome video
   - Added didgeridoo research highlight
-  - Added ability for users to invite others via email
+  - Added ability for users to invite new members via email
+  - Added ability for users to invite their personal care provider via email
 - **Survey Exports**
   - Added a data dictionary export task to export all surveys into CSV format
   - Added a data export task that matches the data dictionary format and is exported into CSV format